backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-socketio>=5.11.0
motor>=3.3.0
pydantic[email]>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# RAG dependencies
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-google-genai>=0.0.1
chromadb>=0.5.0
PyPDF2>=3.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
